feat: training pipeline, engine v2, speaker format v2, subset CLI
Major additions across the full stack:

Training pipeline (Python):
- Fix mel_basis dtype mismatch (float64→float32) in audio.py
- Fix pyloudnorm float64 output causing XPU dtype errors
- VCTKAdapter: support both old (.wav) and 0.92 (.flac) formats
- Add --subset, --segment-min-sec, --num-workers CLI options
  to preprocess, train-teacher, and distill commands
- TMRVCDataset/create_dataloader: subset parameter for quick iteration
- Parallel segmented downloader (scripts/parallel_download.py)
- Verified Teacher Phase 0 training on Intel Arc B570 (XPU)

Rust engine:
- Speaker file v2: metadata JSON + mel thumbnail PNG
- Voice profile creation pipeline with speaker encoder
- WAV reader, mel thumbnail generator, DSP enhancements
- NAM WaveNet integration updates
- Runtime app with voice profile panel

Python export:
- Speaker file v2 read/write with metadata + thumbnail
- ONNX export updates and parity verification enhancements

GUI & VST:
- GUI enrollment page with metadata support
- Finetune/train workers updated for v2 format
- VST plugin parameter and UI updates
- Standalone demo scripts

Docs:
- Updated onnx-contract.md for speaker v2 format
- GUI design document
- Paper-level design document

+++ tmrvc-data/src/tmrvc_data/preprocessing.py
@@ -80,7 +80,7 @@ def normalize_loudness(
         return waveform
 
     normalised = pyln.normalize.loudness(audio_np, current_lufs, target_lufs)
-    return torch.from_numpy(normalised).unsqueeze(0)
+    return torch.from_numpy(normalised).float().unsqueeze(0)
 
 
 # ---------------------------------------------------------------------------
